requirements.txt
amqp==5.2.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.0
Brotli==1.1.0
build==1.2.1
celery==5.4.0
celery-progress==0.4
certifi==2024.7.4
cffi==1.17.0
cfgv==3.4.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==43.0.0
daphne==4.1.2
distlib==0.3.8
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.1.0
django-anymail==11.1
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-health-check==3.18.3
django-hijack==3.6.0
django-otp==1.5.2
django-timezone-field==7.0
django-waffle==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
drf-spectacular==0.27.2
fido2==1.1.3
filelock==3.15.4
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.1
hyperlink==21.0.0
identify==2.6.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
msgpack==1.0.8
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pip-tools==7.4.1
platformdirs==4.2.2
pre-commit==3.8.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pypng==0.20220715.0
pyproject_hooks==1.1.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==7.4.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
ruff==0.6.2
sentry-sdk==2.13.0
service-identity==24.1.0
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
vine==5.1.0
virtualenv==20.26.3
watchfiles==0.23.0
wcwidth==0.2.13
websockets==13.0
wheel==0.44.0
whitenoise==6.7.0
zope.interface==7.0.1
Pillow==10.4.0
